refactor: use isFetching instead of useLoading, outsource getProducts into an api folder, replace messages components with a StateMessage component
