--- src/components/ui/ErrorMessage.tsx
@@ -1,9 +1,0 @@
-export const ErrorMessage = () => {
-  return (
-    <div className="min-h-screen flex justify-center items-center">
-      <p className="text-4xl font-ubuntu text-primary-900">
-        There was an error...
-      </p>
-    </div>
-  );
-};
