--- src/components/ui/HasNoResultsMessage.tsx
@@ -1,9 +1,0 @@
-export const HasNoResultsMessage = () => {
-  return (
-    <div className="min-h-screen flex justify-center items-center">
-      <p className="text-4xl font-ubuntu text-primary-900">
-        No results found...
-      </p>
-    </div>
-  );
-};
