--- src/components/ui/LoadingMessage.tsx
@@ -1,7 +1,0 @@
-export const LoadingMessage = () => {
-  return (
-    <div className="min-h-screen flex justify-center items-center">
-      <p className="text-4xl font-ubuntu text-primary-900">Loading...</p>
-    </div>
-  );
-};
