+++ src/components/ui/index.ts
@@ -1,4 +1,2 @@
 export * from "./Button";
-export * from "./ErrorMessage";
-export * from "./HasNoResultsMessage";
-export * from "./LoadingMessage";
+export * from "./StateMessage";
